Move AST walking to separate visitor-pattern dispatcher
